[Feat] Code resctructure (unfinished)
