+++ Week 06/Lecture 11/Assignment 01/lecture_11/run.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+echo Building the project with Maven...
+mvn clean install && java -jar target/lecture_11-1.0-SNAPSHOT.jar
