+++ Week 06/Lecture 11/Assignment 01/lecture_11/src/main/java/com/example/lecture_11/data/model/Employee.java
@@ -4,8 +4,6 @@ import java.time.LocalDate;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
-import jakarta.persistence.EnumType;
-import jakarta.persistence.Enumerated;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
@@ -38,7 +36,6 @@ public class Employee {
     private String lastName;
 
     @Column(columnDefinition = "enum('M','F')", nullable = false)
-    @Enumerated(EnumType.STRING)
     private String gender;
 
     @Temporal(TemporalType.DATE)
