+++ Week 06/Lecture 11/Assignment 01/lecture_11/src/main/java/com/example/lecture_11/data/model/Salary.java
