+++ Week 06/Lecture 11/Assignment 01/lecture_11/src/main/java/com/example/lecture_11/data/model/composite/DeptEmpId.java
@@ -1,15 +1,15 @@
 package com.example.lecture_11.data.model.composite;
 
 import java.io.Serializable;
-import java.time.LocalDate;
 
+import jakarta.persistence.Embeddable;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 
 @Data
+@Embeddable
 @AllArgsConstructor
-public class TitleId implements Serializable {
+public class DeptEmpId implements Serializable {
     private Integer empNo;
-    private String title;
-    private LocalDate fromDate;
+    private String deptNo;
 }
