+++ Week 06/Lecture 11/Assignment 01/lecture_11/src/main/java/com/example/lecture_11/data/model/composite/DeptManagerId.java
@@ -2,10 +2,12 @@ package com.example.lecture_11.data.model.composite;
 
 import java.io.Serializable;
 
+import jakarta.persistence.Embeddable;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 
 @Data
+@Embeddable
 @AllArgsConstructor
 public class DeptManagerId implements Serializable {
     private Integer empNo;
