+++ Week 06/Lecture 11/Assignment 01/lecture_11/src/main/java/com/example/lecture_11/data/model/composite/SalaryId.java
@@ -3,10 +3,12 @@ package com.example.lecture_11.data.model.composite;
 import java.io.Serializable;
 import java.time.LocalDate;
 
+import jakarta.persistence.Embeddable;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 
 @Data
+@Embeddable
 @AllArgsConstructor
 public class SalaryId implements Serializable {
     private Integer empNo;
