+++ Week 06/Lecture 11/Assignment 01/lecture_11/src/main/java/com/example/lecture_11/data/model/composite/TitleId.java
@@ -1,13 +1,17 @@
 package com.example.lecture_11.data.model.composite;
 
 import java.io.Serializable;
+import java.time.LocalDate;
 
+import jakarta.persistence.Embeddable;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 
 @Data
+@Embeddable
 @AllArgsConstructor
-public class DeptEmpId implements Serializable {
+public class TitleId implements Serializable {
     private Integer empNo;
-    private String deptNo;
+    private String title;
+    private LocalDate fromDate;
 }
