+++ Week 06/Lecture 11/Assignment 01/lecture_11/src/main/resources/application.properties
@@ -1,0 +1,8 @@
+spring.application.name=lecture_11
+
+spring.datasource.url=jdbc:mysql://localhost:3308/week6_lecture11?allowPublicKeyRetrieval=true&useSSL=false
+spring.datasource.username=root
+spring.datasource.password=Michaeleon16606_
+spring.datasource.driver-class-name=com.mysql.jdbc.Driver
+spring.datasource.initialization-mode=always
+spring.jpa.hibernate.ddl-auto=create-drop
