--- Week 06/Lecture 11/lecture_11/src/main/resources/application.properties
@@ -1,1 +1,0 @@
-spring.application.name=lecture_11
